pt-application: Add run once mode
Add a "run-once" mode that checks for the existence of a "run-once" file
in user config directory.

If the file exists, the Phosh Tour skips launching, assuming the tour
has already been completed.

If the file is not present, the application launches as usual and
creates the "run-once" file on startup.

Closes: https://gitlab.gnome.org/World/Phosh/phosh-tour/-/issues/3
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh-tour/-/merge_requests/39>